pyproject.toml
[project]
name = "x-monitor"
version = "0.1.0"
description = "AI Agent for monitoring X.com accounts and generating daily summaries"
requires-python = ">=3.11"
dependencies = [
    "tweepy>=4.14.0",          # Twitter/X API
    "openai>=1.0.0",           # OpenAI API
    "python-telegram-bot>=20.0", # Telegram notifications
    "aiosmtplib>=3.0.0",       # Async email
    "apscheduler>=3.10.0",     # Task scheduling
    "pydantic>=2.0.0",         # Data validation
    "pydantic-settings>=2.0.0", # Settings management
    "httpx>=0.25.0",           # HTTP client
    "python-dotenv>=1.0.0",    # Environment variables
    "loguru>=0.7.0",           # Logging
    "aiosqlite>=0.19.0",       # Async SQLite for storage
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
x-monitor = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
